utils and model refactor
