+++ src/utils/pagination.py
@@ -1,15 +1,5 @@
-import psycopg2
-from .config import settings
 from fastapi import Request
 from urllib.parse import urlencode
-
-def get_connection() -> psycopg2.extensions.connection:
-    db_url = (
-        settings.TEST_DATABASE_URL
-        if settings.ENV == "test"
-        else settings.DATABASE_URL
-    )
-    return psycopg2.connect(db_url)
 
 def paginate_results(items: list, total: int, request: Request, limit: int, offset: int):
     base_url = str(request.url).split("?")[0]
